+++ src/lib/viewers/controls/findbar/FindBarToggle.tsx
@@ -14,6 +14,7 @@ export default function FindBarToggle({ onFindBarToggle }: Props): JSX.Element |
     return (
         <button
             className="bp-FindBarToggle"
+            data-resin-target="findBar"
             onClick={({ target }): void => onFindBarToggle(target)}
             title={__('toggle_findbar')}
             type="button"